Refactor testing parameter parsing
