+++ testing/adios2/engine/sst/TestSstWriterFails.cpp
@@ -11,6 +11,8 @@
 
 #include <adios2.h>
 
+#include "../ParamsHelpers.h"
+
 #include <gtest/gtest.h>
 
 class SstWriteFails : public ::testing::Test
@@ -24,39 +26,6 @@ std::string fname = "ADIOS2SstFAIL";
 
 int CompressSz = 0;
 int CompressZfp = 0;
-
-static std::string Trim(std::string &str)
-{
-    size_t first = str.find_first_not_of(' ');
-    size_t last = str.find_last_not_of(' ');
-    return str.substr(first, (last - first + 1));
-}
-
-/*
- * Engine parameters spec is a poor-man's JSON.  name:value pairs are separated
- * by commas.  White space is trimmed off front and back.  No quotes or anything
- * fancy allowed.
- */
-static adios2::Params ParseEngineParams(std::string Input)
-{
-    std::istringstream ss(Input);
-    std::string Param;
-    adios2::Params Ret = {};
-
-    while (std::getline(ss, Param, ','))
-    {
-        std::istringstream ss2(Param);
-        std::string ParamName;
-        std::string ParamValue;
-        std::getline(ss2, ParamName, ':');
-        if (!std::getline(ss2, ParamValue, ':'))
-        {
-            throw std::invalid_argument("Engine parameter \"" + Param + "\" missing value");
-        }
-        Ret[Trim(ParamName)] = Trim(ParamValue);
-    }
-    return Ret;
-}
 
 // ADIOS2 SST write
 TEST_F(SstWriteFails, InvalidPut)
