+++ testing/adios2/performance/metadata/PerfMetaData.cpp
@@ -15,6 +15,8 @@
 
 #include <adios2.h>
 
+#include "../../engine/ParamsHelpers.h"
+
 #ifndef _WIN32
 #include "strings.h"
 #else
@@ -56,39 +58,6 @@ typedef enum
 } TestModeEnum;
 
 TestModeEnum TestMode = WriterConsolidation;
-
-static std::string Trim(std::string &str)
-{
-    size_t first = str.find_first_not_of(' ');
-    size_t last = str.find_last_not_of(' ');
-    return str.substr(first, (last - first + 1));
-}
-
-/*
- * Engine parameters spec is a poor-man's JSON.  name:value pairs are separated
- * by equal.  White space is trimmed off front and back.  No quotes or anything
- * fancy allowed.
- */
-static adios2::Params ParseEngineParams(std::string Input)
-{
-    std::istringstream ss(Input);
-    std::string Param;
-    adios2::Params Ret = {};
-
-    while (std::getline(ss, Param, ','))
-    {
-        std::istringstream ss2(Param);
-        std::string ParamName;
-        std::string ParamValue;
-        std::getline(ss2, ParamName, '=');
-        if (!std::getline(ss2, ParamValue, '='))
-        {
-            throw std::invalid_argument("Engine parameter \"" + Param + "\" missing value");
-        }
-        Ret[Trim(ParamName)] = Trim(ParamValue);
-    }
-    return Ret;
-}
 
 static void Usage()
 {
